# Skip sstables that are entirely covered by a range tombstone