--- a/db/forward_iterator.cc
+++ b/db/forward_iterator.cc
@@ -38,6 +38,7 @@ class ForwardLevelIterator : public InternalIterator {
       : cfd_(cfd),
         read_options_(read_options),
         files_(files),
+        range_del_agg_(cfd->internal_comparator(), {} /* snapshots */),
         valid_(false),
         file_index_(std::numeric_limits<uint32_t>::max()),
         file_iter_(nullptr),
@@ -71,16 +72,14 @@ class ForwardLevelIterator : public InternalIterator {
       delete file_iter_;
     }
 
-    RangeDelAggregator range_del_agg(
-        cfd_->internal_comparator(), {} /* snapshots */);
     file_iter_ = cfd_->table_cache()->NewIterator(
         read_options_, *(cfd_->soptions()), cfd_->internal_comparator(),
         *files_[file_index_],
-        read_options_.ignore_range_deletions ? nullptr : &range_del_agg,
+        read_options_.ignore_range_deletions ? nullptr : &range_del_agg_,
         prefix_extractor_, nullptr /* table_reader_ptr */, nullptr, false);
     file_iter_->SetPinnedItersMgr(pinned_iters_mgr_);
     valid_ = false;
-    if (!range_del_agg.IsEmpty()) {
+    if (!range_del_agg_.IsEmpty()) {
       status_ = Status::NotSupported(
           "Range tombstones unsupported with ForwardIterator");
     }
@@ -184,6 +183,7 @@ class ForwardLevelIterator : public InternalIterator {
   const ColumnFamilyData* const cfd_;
   const ReadOptions& read_options_;
   const std::vector<FileMetaData*>& files_;
+  RangeDelAggregator range_del_agg_;
 
   bool valid_;
   uint32_t file_index_;
@@ -202,6 +202,7 @@ ForwardIterator::ForwardIterator(DBImpl* db, const ReadOptions& read_options,
       prefix_extractor_(current_sv->mutable_cf_options.prefix_extractor.get()),
       user_comparator_(cfd->user_comparator()),
       immutable_min_heap_(MinIterComparator(&cfd_->internal_comparator())),
+      range_del_agg_(cfd->internal_comparator(), {} /* snapshots */),
       sv_(current_sv),
       mutable_iter_(nullptr),
       current_(nullptr),
@@ -608,16 +609,14 @@ void ForwardIterator::RebuildIterators(bool refresh_sv) {
     // New
     sv_ = cfd_->GetReferencedSuperVersion(&(db_->mutex_));
   }
-  RangeDelAggregator range_del_agg(
-      cfd_->internal_comparator(), {} /* snapshots */);
   mutable_iter_ = sv_->mem->NewIterator(read_options_, &arena_);
   sv_->imm->AddIterators(read_options_, &imm_iters_, &arena_);
   if (!read_options_.ignore_range_deletions) {
     std::unique_ptr<InternalIterator> range_del_iter(
         sv_->mem->NewRangeTombstoneIterator(read_options_));
-    range_del_agg.AddTombstones(std::move(range_del_iter));
+    range_del_agg_.AddTombstones(std::move(range_del_iter));
     sv_->imm->AddRangeTombstoneIterators(read_options_, &arena_,
-                                         &range_del_agg);
+                                         &range_del_agg_);
   }
   has_iter_trimmed_for_upper_bound_ = false;
 
@@ -636,7 +635,7 @@ void ForwardIterator::RebuildIterators(bool refresh_sv) {
     }
     l0_iters_.push_back(cfd_->table_cache()->NewIterator(
         read_options_, *cfd_->soptions(), cfd_->internal_comparator(), *l0,
-        read_options_.ignore_range_deletions ? nullptr : &range_del_agg,
+        read_options_.ignore_range_deletions ? nullptr : &range_del_agg_,
         sv_->mutable_cf_options.prefix_extractor.get()));
   }
   BuildLevelIterators(vstorage);
@@ -644,7 +643,7 @@ void ForwardIterator::RebuildIterators(bool refresh_sv) {
   is_prev_set_ = false;
 
   UpdateChildrenPinnedItersMgr();
-  if (!range_del_agg.IsEmpty()) {
+  if (!range_del_agg_.IsEmpty()) {
     status_ = Status::NotSupported(
         "Range tombstones unsupported with ForwardIterator");
     valid_ = false;
@@ -666,14 +665,12 @@ void ForwardIterator::RenewIterators() {
 
   mutable_iter_ = svnew->mem->NewIterator(read_options_, &arena_);
   svnew->imm->AddIterators(read_options_, &imm_iters_, &arena_);
-  RangeDelAggregator range_del_agg(
-      cfd_->internal_comparator(), {} /* snapshots */);
   if (!read_options_.ignore_range_deletions) {
     std::unique_ptr<InternalIterator> range_del_iter(
         svnew->mem->NewRangeTombstoneIterator(read_options_));
-    range_del_agg.AddTombstones(std::move(range_del_iter));
+    range_del_agg_.AddTombstones(std::move(range_del_iter));
     svnew->imm->AddRangeTombstoneIterators(read_options_, &arena_,
-                                           &range_del_agg);
+                                           &range_del_agg_);
   }
 
   const auto* vstorage = sv_->current->storage_info();
@@ -707,7 +704,7 @@ void ForwardIterator::RenewIterators() {
     l0_iters_new.push_back(cfd_->table_cache()->NewIterator(
         read_options_, *cfd_->soptions(), cfd_->internal_comparator(),
         *l0_files_new[inew],
-        read_options_.ignore_range_deletions ? nullptr : &range_del_agg,
+        read_options_.ignore_range_deletions ? nullptr : &range_del_agg_,
         svnew->mutable_cf_options.prefix_extractor.get()));
   }
 
@@ -728,7 +725,7 @@ void ForwardIterator::RenewIterators() {
   sv_ = svnew;
 
   UpdateChildrenPinnedItersMgr();
-  if (!range_del_agg.IsEmpty()) {
+  if (!range_del_agg_.IsEmpty()) {
     status_ = Status::NotSupported(
         "Range tombstones unsupported with ForwardIterator");
     valid_ = false;