--- a/db/range_del_aggregator.cc
+++ b/db/range_del_aggregator.cc
@@ -80,14 +80,14 @@ class UncollapsedRangeDelMap : public RangeDelMap {
     return false;
   }
 
-  RangeTombstone GetTombstone(const Slice& user_key, SequenceNumber seqno) {
-    // Unimplemented because the only client of this method, iteration, uses
-    // collapsed maps.
+  std::pair<RangePtr,SequenceNumber> GetTombstone(
+      const Slice& user_key, SequenceNumber seqno) {
+    // Unimplemented, though the lack of implementation only affects
+    // performance (not correctness) for sstable ingestion. Normal
+    // read operations use a CollapsedRangeDelMap.
     (void)user_key;
     (void)seqno;
-    fprintf(stderr, "UncollapsedRangeDelMap::GetTombstone unimplemented");
-    abort();
-    return RangeTombstone(Slice(), Slice(), 0);
+    return std::make_pair(RangePtr(), 0);
   }
 
   bool IsRangeOverlapped(const Slice& start, const Slice& end) {
@@ -193,7 +193,8 @@ class CollapsedRangeDelMap : public RangeDelMap {
   const Comparator* ucmp_;
 
  public:
-  CollapsedRangeDelMap(const Comparator* ucmp) : ucmp_(ucmp) {
+  CollapsedRangeDelMap(const Comparator* ucmp)
+      : rep_(stl_wrappers::LessOfComparator(ucmp)), ucmp_(ucmp) {
     InvalidatePosition();
   }
 
@@ -270,11 +271,12 @@ class CollapsedRangeDelMap : public RangeDelMap {
     }
     --iter;
     if (ucmp_->Compare(parsed_start.user_key, iter->first) < 0) {
+      assert(false);
       return false;
     }
-    // Loop looking for a tombstone that is older than the range
-    // sequence number, or we determine that our range is completely
-    // covered by newer tombstones.
+    // Loop looking for a tombstone that is older than the range sequence
+    // number, or we determine that our range is completely covered by newer
+    // tombstones.
     for (; iter != rep_.end(); ++iter) {
       if (ucmp_->Compare(parsed_end.user_key, iter->first) < 0) {
         return true;
@@ -287,20 +289,24 @@ class CollapsedRangeDelMap : public RangeDelMap {
     return false;
   }
 
-  RangeTombstone GetTombstone(const Slice& user_key, SequenceNumber seqno) {
+  std::pair<RangePtr,SequenceNumber> GetTombstone(
+      const Slice& user_key, SequenceNumber seqno) {
     auto iter = rep_.upper_bound(user_key);
     if (iter == rep_.begin()) {
       // before start of deletion intervals
-      return RangeTombstone(Slice(), iter->first, 0);
+      return std::make_pair(RangePtr(nullptr, &iter->first), 0);
     }
     auto prev = iter;
     --prev;
     if (iter == rep_.end()) {
       // after end of deletion intervals
-      return RangeTombstone(prev->first, Slice(), 0);
+      return std::make_pair(RangePtr(&prev->first, nullptr), 0);
     }
-    return RangeTombstone(prev->first, iter->first,
-                          prev->second >= seqno ? prev->second : 0);
+    // Note that a range tombstone does not cover a key at the same sequence
+    // number. This can occur in an sstable that has been ingested where all
+    // of the entries have the same sequence number.
+    return std::make_pair(RangePtr(&prev->first, &iter->first),
+                          prev->second > seqno ? prev->second : 0);
   }
 
   bool IsRangeOverlapped(const Slice&, const Slice&) {
@@ -420,7 +426,12 @@ class CollapsedRangeDelMap : public RangeDelMap {
     }
   }
 
-  size_t Size() const { return rep_.size() - 1; }
+  size_t Size() const {
+    if (rep_.empty()) {
+      return 0;
+    }
+    return rep_.size() - 1;
+  }
 
   void InvalidatePosition() { iter_ = rep_.end(); }
 
@@ -500,14 +511,14 @@ bool RangeDelAggregator::ShouldDeleteRange(
   return tombstone_map.ShouldDeleteRange(start, end, seqno);
 }
 
-RangeTombstone RangeDelAggregator::GetTombstone(const Slice& user_key,
-                                                SequenceNumber seqno) {
+std::pair<RangePtr,SequenceNumber> RangeDelAggregator::GetTombstone(
+    const Slice& user_key, SequenceNumber seqno) {
   if (rep_ == nullptr) {
-    return RangeTombstone(Slice(), Slice(), 0);
+    return std::make_pair(RangePtr(), 0);
   }
   auto& tombstone_map = GetRangeDelMap(seqno);
   if (tombstone_map.IsEmpty()) {
-    return RangeTombstone(Slice(), Slice(), 0);
+    return std::make_pair(RangePtr(), 0);
   }
   return tombstone_map.GetTombstone(user_key, seqno);
 }