--- a/db/range_del_aggregator.h
+++ b/db/range_del_aggregator.h
@@ -62,6 +62,10 @@ class RangeDelMap {
 
   virtual bool ShouldDelete(const ParsedInternalKey& parsed,
                             RangeDelPositioningMode mode) = 0;
+  virtual bool ShouldDeleteRange(const Slice& start, const Slice& end,
+                                 SequenceNumber seqno) = 0;
+  virtual RangeTombstone GetTombstone(const Slice& user_key,
+                                      SequenceNumber seqno) = 0;  
   virtual bool IsRangeOverlapped(const Slice& start, const Slice& end) = 0;
   virtual void InvalidatePosition() = 0;
 
@@ -134,6 +138,13 @@ class RangeDelAggregator {
   // the largest_key metadata for an sstable.
   bool ShouldDeleteRange(const Slice& start, const Slice& end,
                          SequenceNumber seqno);
+
+  // Get the range tombstone at the specified user_key and sequence number. A
+  // valid tombstone is always returned, though it may cover an empty range of
+  // keys or the sequence number may be 0 to indicate that no tombstone covers
+  // the specified range of keys.
+  RangeTombstone GetTombstone(const Slice& user_key,
+                              SequenceNumber seqno);
 
   // Checks whether range deletions cover any keys between `start` and `end`,
   // inclusive.