--- a/db/range_del_aggregator.h
+++ b/db/range_del_aggregator.h
@@ -128,6 +128,13 @@ class RangeDelAggregator {
   bool ShouldDeleteImpl(const Slice& internal_key,
                         RangeDelPositioningMode mode);
 
+  // Return true if one or more tombstones that are newer than seqno fully
+  // cover the range [start,end] (both endpoints are inclusive). Beware the
+  // inclusive endpoint which differs from most other key ranges, but matches
+  // the largest_key metadata for an sstable.
+  bool ShouldDeleteRange(const Slice& start, const Slice& end,
+                         SequenceNumber seqno);
+
   // Checks whether range deletions cover any keys between `start` and `end`,
   // inclusive.
   //