--- a/db/version_set.cc
+++ b/db/version_set.cc
@@ -544,8 +544,8 @@ class LevelIterator final : public InternalIterator {
     assert(file_index_ < flevel_->num_files);
     auto file_meta = flevel_->files[file_index_];
     // Check to see if every key in the sstable is covered by a range
-    // tombstone. SkipEmptyFile{Forward,Backward} will take care of
-    // skipping over an "empty" file if we return null.
+    // tombstone. SkipEmptyFile{Forward,Backward} will take care of skipping
+    // over an "empty" file if we return null.
     if (range_del_agg_->ShouldDeleteRange(file_meta.smallest_key, file_meta.largest_key,
                                           file_meta.file_metadata->largest_seqno)) {
       return nullptr;