--- a/table/cuckoo_table_reader.h
+++ b/table/cuckoo_table_reader.h
@@ -48,6 +48,8 @@ class CuckooTableReader: public TableReader {
 
   InternalIterator* NewIterator(const ReadOptions&,
                                 const SliceTransform* prefix_extractor,
+                                RangeDelAggregator* range_del_agg = nullptr,
+                                const FileMetaData* file_meta = nullptr,
                                 Arena* arena = nullptr,
                                 bool skip_filters = false,
                                 bool for_compaction = false) override;