During iteration skip sstables that are entirely covered by a range
tombstone. The check to do this is loose: we check that an sstable's
[smallest_key,largest_key) at largest_seqno is covered by a range
tombstone.

This PR still needs tests. I'm having trouble running them on by Mac
laptop. I wanted to get this out for review to see if there is anything
fundamentally wrong with this approach.