@petermattis 2018-06-08
@ajkr The unwritten test I want to write involves creating an sstable with N keys (say 10k), adding a range tombstone in the memtable that covers these keys and then creating an iterator and seeking to a key near the start of the sstable. Before this patch, the Seek would iterate through all of the keys in the sstable calling `RangeDelAggregator::ShouldDelete`. After this change iteration completely skips the covered sstable.

This PR suggests another enhancement: `LevelIterator` should take into account range tombstones during `Seek` and `SeekForPrev`. I think this can be accomplished by consulting `RangeDelAggregator` and utilizing the current file's `largest_seqno`. Given the seek key and a sequence number, what is the next (or previous) key that is not deleted. We'd then bound that returned key by the sstable's key range.