@petermattis 2018-06-08
I figured out my test problem. All tests pass with this PR and I've added testing of the new `RangeDelAggregator::ShouldDeleteRange` method. I'm not quite sure how to test that the `LevelIterator::InitFileIterator` is actually performing this optimization. Suggestions?