@petermattis 2018-06-08
Here are benchmark numbers. The benchmarks create an sstable with 10, 1000, and 100000 entries, creates a range tombstone covering all of those entries and then seeks to the first entry (which will be deleted). 

```
name                                        old time/op  new time/op  delta
RocksDBDeleteRangeIterate/entries=10-8      6.65µs ± 2%  4.63µs ± 1%  -30.33%  (p=0.000 n=10+10)
RocksDBDeleteRangeIterate/entries=1000-8    84.6µs ± 2%   4.7µs ± 3%  -94.41%  (p=0.000 n=10+10)
RocksDBDeleteRangeIterate/entries=100000-8  7.60ms ± 4%  0.00ms ± 2%  -99.94%  (p=0.000 n=10+10)
```

I have an idea for how to adjust the implementation here so that we can take advantage of skipping over an entire range of deleted keys in `LevelIterator::{Seek,SeekForPrev,Next,Prev}` even when the range tombstone only covers part of the sstable.