@petermattis 2018-06-09
I've added a second commit which mentions the other idea I mentioned above: skipping over ranges of keys in `LevelIterator` that are covered by a range tombstone. This isn't quite as effective as the sstable skipping optimization, but applies in a lot more cases. I doubt there is enough testing in this PR yet, but all the existing tests are passing.