@benesch 2018-06-15
@petermattis want me to shepard this in your absence?

On Fri, Jun 15, 2018 at 6:19 PM Facebook Community Bot <
notifications@github.com> wrote:

> @petermattis <https://github.com/petermattis> has updated the pull
> request.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/facebook/rocksdb/pull/3969#issuecomment-397756420>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AA15IOnU0vpkZ-_3cfXbUF3pKWGF9Fqyks5t9DLrgaJpZM4UgJWQ>
> .
>