@benesch 2018-06-21
After pondering this patch for a bit, it seems like the logic for seeking to the end of a tombstone should live in the BlockBasedTableIterator so that it applies to files in all levels, including L0. I also suspect that it may be less redundant/awkward there.