@petermattis 2018-07-09
@ajkr I reworked this PR to implement the tombstone skipping in `BlockBasedTableIterator`. This followed a very similar structure to the code that I had previously implemented in `LevelIterator`. Still to do is to add more dedicated test cases. Do you have an opinion on where those tests should go?