@benesch 2018-09-12
Just so we don't forget: this PR should incorporate cockroachdb/rocksdb#12, cockroachdb/rocksdb#13, and cockroachdb/rocksdb#15 before it merges.